2-get_bit.c fixed

+++ bit_manipulation/2-get_bit.c
@@ -9,7 +9,7 @@
 
 int get_bit(unsigned long int n, unsigned int index)
 {
-	int b;
+	unsigned long int b;
 	int range = 63 - index;
